Added Javadoc to all other files for #29
