+++ src/main/java/cc/bluebits/hongtaiyang/block/RootedSculkBlock.java
@@ -2,8 +2,16 @@ package cc.bluebits.hongtaiyang.block;
 
 import net.minecraft.world.level.block.Block;
 
+/**
+ * A rooted variant of the sculk block
+ * @see net.minecraft.world.level.block.SculkBlock
+ */
 @SuppressWarnings("unused")
 public class RootedSculkBlock extends Block {
+	/**
+	 * Constructs a {@code RootedSculkBlock}
+	 * @param pProperties The properties of the block
+	 */
 	public RootedSculkBlock(Properties pProperties) {
 		super(pProperties);
 	}
