--- src/main/java/cc/bluebits/hongtaiyang/block/base/ModFlammableRotatedPillarBlock.java
@@ -1,44 +1,0 @@
-package cc.bluebits.hongtaiyang.block.base;
-
-import cc.bluebits.hongtaiyang.registries.block.ModBlocks;
-import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.world.item.AxeItem;
-import net.minecraft.world.item.context.UseOnContext;
-import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.block.RotatedPillarBlock;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraftforge.common.ToolAction;
-import org.jetbrains.annotations.Nullable;
-
-public class ModFlammableRotatedPillarBlock extends RotatedPillarBlock {
-	public ModFlammableRotatedPillarBlock(Properties pProperties) {
-		super(pProperties);
-	}
-
-	@Override
-	public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {
-		return true;
-	}
-
-	@Override
-	public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {
-		return 5; // General value for logs
-	}
-
-	@Override
-	public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {
-		return 5; // General value for logs
-	}
-
-	@Override
-	public @Nullable BlockState getToolModifiedState(BlockState state, UseOnContext context, ToolAction toolAction, boolean simulate) {
-		if (context.getItemInHand().getItem() instanceof AxeItem) {
-			if (state.is(ModBlocks.DARKDWELLER_BUNDLE.get())) {
-				return ModBlocks.STRIPPED_DARKDWELLER_BUNDLE.get().defaultBlockState().setValue(AXIS, state.getValue(AXIS));
-			}
-		}
-
-		return super.getToolModifiedState(state, context, toolAction, simulate);
-	}
-}
