+++ src/main/java/cc/bluebits/hongtaiyang/block/base/ModThinPillarFruitBlock.java
@@ -26,9 +26,19 @@ import org.jetbrains.annotations.NotNull;
 import javax.annotation.Nullable;
 import java.util.function.Supplier;
 
+/**
+ * A fruit block that grows on a thin pillar
+ * @see ModModularPillarBlock
+ */
 @SuppressWarnings("deprecation")
 public class ModThinPillarFruitBlock extends HorizontalDirectionalBlock implements BonemealableBlock {
+	/**
+	 * The maximum age of the fruit
+	 */
 	public static final int MAX_AGE = 2;
+	/**
+	 * The block property for the age of the fruit
+	 */
 	public static final IntegerProperty AGE = BlockStateProperties.AGE_2;
 
 
@@ -45,6 +55,11 @@ public class ModThinPillarFruitBlock extends HorizontalDirectionalBlock implemen
 	}
 
 
+	/**
+	 * Constructs a {@code ModThinPillarFruitBlock}
+	 * @param pProperties The block properties, passed to the super constructor
+	 * @param survivesOn The block that this fruit can survive on
+	 */
 	public ModThinPillarFruitBlock(Properties pProperties, Supplier<Block> survivesOn) {
 		super(pProperties);
 		this.SURVIVES_ON = survivesOn;
@@ -55,6 +70,12 @@ public class ModThinPillarFruitBlock extends HorizontalDirectionalBlock implemen
 	}
 
 
+	/**
+	 * Gets the index of the shape based on the age and facing of the fruit
+	 * @param age The age of the fruit 
+	 * @param facing The direction the fruit is facing
+	 * @return the index of the shape in the SHAPES array
+	 */
 	public static int getShapeIndex(int age, Direction facing) {
 		int nShapesPerAge = (int) Direction.Plane.HORIZONTAL.stream().count();
 		int ageOffset = nShapesPerAge * age;
