+++ src/main/java/cc/bluebits/hongtaiyang/registries/item/ModFoods.java
@@ -4,6 +4,10 @@ import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.food.FoodProperties;
 
+/**
+ * Class for registering foods
+ */
+@SuppressWarnings("MissingJavadoc")
 public class ModFoods {
 	public static final FoodProperties DWELLBERRY = new FoodProperties.Builder()
 			.nutrition(2)
