+++ src/main/java/cc/bluebits/hongtaiyang/registries/world/feature/ModFeatures.java
@@ -9,9 +9,20 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
+/**
+ * Class for registering features
+ */
+@SuppressWarnings("MissingJavadoc")
 public class ModFeatures {
+	/**
+	 * The deferred register for features
+	 */
 	public static final DeferredRegister<Feature<?>> FEATURES = DeferredRegister.create(ForgeRegistries.FEATURES, HongTaiyang.MOD_ID);
 
+	/**
+	 * Method for registering the features
+	 * @param eventBus The event bus to register to
+	 */
 	public static void register(IEventBus eventBus) {
 		FEATURES.register(eventBus);
 	}
