+++ src/main/java/cc/bluebits/hongtaiyang/registries/world/feature/tree/ModFoliagePlacers.java
@@ -8,9 +8,20 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
+/**
+ * Class for registering foliage placers
+ */
+@SuppressWarnings("MissingJavadoc")
 public class ModFoliagePlacers {
+	/**
+	 * The deferred register for foliage placers
+	 */
 	public static final DeferredRegister<FoliagePlacerType<?>> FOLIAGE_PLACERS = DeferredRegister.create(ForgeRegistries.FOLIAGE_PLACER_TYPES, HongTaiyang.MOD_ID);
 
+	/**
+	 * Method for registering the foliage placers
+	 * @param eventBus The event bus to register to
+	 */
 	public static void register(IEventBus eventBus) {
 		FOLIAGE_PLACERS.register(eventBus);
 	}
