+++ src/main/java/cc/bluebits/hongtaiyang/registries/world/feature/tree/ModTreeDecorators.java
@@ -8,9 +8,20 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.RegistryObject;
 
+/**
+ * Class for registering tree decorators
+ */
+@SuppressWarnings("MissingJavadoc")
 public class ModTreeDecorators {
+	/**
+	 * The deferred register for tree decorators
+	 */
 	public static final DeferredRegister<TreeDecoratorType<?>> TREE_DECORATORS = DeferredRegister.create(Registries.TREE_DECORATOR_TYPE, HongTaiyang.MOD_ID);
 
+	/**
+	 * Method for registering the tree decorators
+	 * @param eventBus The event bus
+	 */
 	public static void register(IEventBus eventBus) {
 		TREE_DECORATORS.register(eventBus);
 	}
